added token authentication

+++ controllers/books.js
@@ -1,8 +1,18 @@
 const booksRouter = require('express').Router()
 const Book = require('../models/book')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+
+const getTokenFrom = request => {
+  const authorization = request.get('authorization')
+  if (authorization && authorization.startsWith('Bearer ')) {
+    return authorization.replace('Bearer ', '')
+  }
+  return null
+}
 
 booksRouter.get('/', async (request, response) => {
-  const books = await Book.find({})
+  const books = await Book.find({}).populate('user', { username: 1, email: 1 })
 
   response.json(books)
 })
@@ -19,28 +29,54 @@ booksRouter.get('/:id', async (request, response) => {
 })
 
 booksRouter.post('/', async (request, response) => {
-  const book = new Book(request.body)
+  const { body } = request;
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
 
-  if (!book.key) {
-    return response.status(400).send({ error: 'name or number not present' })
-  }
-
-  if (!book.userInfo) {
-    book.userInfo = {}
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })  
   }
 
-  if (!book.userInfo.status) {
-    book.userInfo.status = 'reading'
+  const user = await User.findById(decodedToken.id)
+
+  if (!body.key) {
+    return response.status(400).end()
   }
-  if (!book.userInfo.notes) {
-    book.userInfo.notes = ''
+
+  if (!body.userInfo) {
+    body.userInfo = {}
   }
+
+  if (!body.userInfo.status) {
+    body.userInfo.status = 'reading'
+  }
+  if (!body.userInfo.notes) {
+    body.userInfo.notes = ''
+  }
+
+  const book = new Book({
+    key: body.key,
+    userInfo: {
+      notes: body.userInfo.notes,
+      status: body.userInfo.status
+    },
+    user: user.id
+  })
 
   const result = await book.save()
+
+  user.books = user.books.concat(result._id)
+  await user.save()
+
   response.status(201).json(result)
 })
 
 booksRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })  
+  }
+
   await Book.findByIdAndDelete(request.params.id)
   response.status(204).end()
 })
@@ -48,6 +84,13 @@ booksRouter.delete('/:id', async (request, response) => {
 booksRouter.put('/:id', async (request, response) => {
   const oldBook = await Book.findById(request.params.id)
   const body = request.body
+
+  const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET)
+  const user = await User.findById(decodedToken.id)
+  
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' })  
+  }
 
   if (!body.key) {
     body.key = oldBook.key
@@ -62,7 +105,8 @@ booksRouter.put('/:id', async (request, response) => {
       userInfo: {
         notes: body.userInfo.notes,
         status: body.userInfo.status
-      }
+      },
+      user: user
     }
     await Book.findByIdAndUpdate(request.params.id, newBook, { new: true })
     response.status(200).json(newBook)
